Renamed IntegrateHead to HeadHandler, added RemoveHead, made static, extracted test function
 - test function now in HeadHandler_RigTest: H to add, J to remove head in scenes/test/rigtest_xexano
