+++ Assets/scripts/animation/HeadHandler.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 533401d59758af34eb937ea4efe6b0f3
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
