+++ Assets/scripts/animation/HeadHandler_RigTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c35f0d25f77728f4e9abd92eff9a78b1
+guid: d9cca9fd3a9769446a3e93e5a59dcdce
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
