--- Assets/scripts/animation/IntegrateHead.cs
@@ -1,81 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class IntegrateHead : MonoBehaviour {
-	
-	// for interactive testing...
-	void Update () {
-		if(Input.GetKeyDown(KeyCode.H)) {
-			GameObject body = GameObject.Find("RigtestBody"); // already in scene
-			AddHead(body, "test/RigtestHead", "Neck");
-		}
-	}
-
-	/**
-	 * Adds an instance of the named headObject to the body and integrates the rig.
-	 */
-	public void AddHead(GameObject body, string headObjectID, string rigRootID="Rig")
-	{
-		GameObject head = (GameObject)Instantiate(Resources.Load(headObjectID)); // from prefab resource, not in scene yet
-		if(head == null) {
-			print("[AddHead] warning: Could not find object "+headObjectID);
-			return;
-		}
-
-		// find rig roots in head and body
-		Transform rigRootHead = null;
-		foreach(Transform transform in head.transform.GetComponentsInChildren<Transform>()) {
-			if(transform.name == rigRootID) {
-				rigRootHead = transform;
-			}
-		}
-		if (rigRootHead == null) {
-			print ("[AddHead] Error: Rig root transform in head not found!");
-			return;
-		}
-
-		Transform rigRootBody = null;
-		foreach(Transform transform in body.transform.GetComponentsInChildren<Transform>()) {
-			if(transform.name == rigRootID) {
-				rigRootBody = transform;
-			}
-		}
-		if (rigRootBody == null) {
-			print ("[AddHead] Error: Rig root transform in body not found!");
-			return;
-		}
-
-		Transform[] bonesInBody = rigRootBody.GetComponentsInChildren<Transform>();
-
-		// reparent bones
-		foreach (Transform boneFromHead in rigRootHead.GetComponentsInChildren<Transform>())
-		{
-			Transform boneFromParent = FindChildTransform(body.transform, bonesInBody, boneFromHead.name);
-			if(boneFromParent != null) {
-				print ("[AddHead] Reparenting bone "+boneFromHead.name);
-
-				boneFromHead.transform.parent = boneFromParent.transform;
-				boneFromHead.transform.localPosition = Vector3.zero;
-				boneFromHead.transform.localRotation = Quaternion.identity;
-
-				boneFromHead.name = boneFromHead.name + "_fromHead";
-			}
-		}
-
-		// attach head object and clear transform (has to fit from model!)
-		head.transform.parent = body.transform;
-		head.transform.localPosition = Vector3.zero;
-		head.transform.localRotation = Quaternion.identity;
-	}
-
-	private Transform FindChildTransform(Transform parent, Transform[] bonesInBody, string name)
-	{
-		foreach (Transform child in bonesInBody) {
-			if(child.name == name) {
-				return child;
-			}
-		}
-
-		return null;
-	}
-}
